refactor: move components out of app directory
