+++ apps/nextjs/src/app/review/page.tsx
@@ -1,4 +1,4 @@
-import { CreateReviewForm } from "../_components/reviews";
+import { CreateReviewForm } from "../../components/reviews";
 
 export default async function Review() {
   return (
